add versions of WinWindow that try out the possibilities for calling win32 functions.
adding version of WinWindow library that uses FFI, due to problems with DL on ruby 1.9 (different and incompatible API) and 1.8.6p383 (segfaults).
unfortunately the FFI version also segfaults on the current FFI gem. but there is a patch in the FFI's master that fixes this.

adding WinWindow version for Win32API as it seems to be the only one that doesn't segfault.
move all of the stuff specific to Win32API into its own module in such a way that I can change to FFI or DL by just changing out the one module.

moving the DL version to a filename reflecting that (DL version using the ruby 1.8 version of DL's api, as the 1.9 DL api is totally different)